v1.8 : clics feutrés, ergonomie petits écrans, corrections
- Clics : « toc » feutré façon clavier iOS au lieu des bips, seulement sur les
  changements de sélection et +/−, anti-rafale, réglage « Clics de l'interface »
- Audit automatique des écrans (320 à 430 px, clair/sombre, Chromium/WebKit)
- Corrections : steppers qui débordaient, titre de barre sous les boutons,
  pastilles et titres tronqués, « Modifier » façon iOS, dates courtes,
  pictos du planning, cibles tactiles 44 px, textes flottants lisibles,
  toast de record en doublon supprimé

+++ forge_projet/src/view_history.js
@@ -43,7 +43,7 @@ function renderHistory(){
       return `<button class="row tap stagger" style="--i:${Math.min(i++,12)}" data-a="openSessionDetail" data-id="${s.id}">
         ${sessionIcon(s)}
         <div class="grow"><div class="t">${esc(sessionTitle(s))}${prs?` <span class="pr-badge">💥 ${prs}</span>`:""}</div>
-        <div class="s">${esc(fmtDate(s.date,"long"))} · ${sessionSetCount(s)} séries${sessionVolume(s)?" · "+fmtKg(sessionVolume(s)):""}</div></div>
+        <div class="s">${esc(fmtDate(s.date,"short"))} · ${sessionSetCount(s)} séries${sessionVolume(s)?" · "+fmtKg(sessionVolume(s)):""}</div></div>
         <div class="val">${s.durationSec?fmtDuration(s.durationSec):""}</div><span class="chev">${icon("chev")}</span>
       </button>`;
     }).join("");
